requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
tensorflow>=2.10.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
Werkzeug>=3.0.0
